+++ src/main/java/blackjack/card/Number.java
@@ -1,0 +1,28 @@
+package blackjack.card;
+
+public enum Number {
+
+    ACE(1),
+    TWO(2),
+    THREE(3),
+    FOUR(4),
+    FIVE(5),
+    SIX(6),
+    SEVEN(7),
+    EIGHT(8),
+    NINE(9),
+    TEN(10),
+    JACK(10),
+    QUEEN(10),
+    KING(10);
+
+    private final int score;
+
+    Number(int score) {
+        this.score = score;
+    }
+
+    int getScore() {
+        return this.score;
+    }
+}
