+++ src/main/java/blackjack/card/Shape.java
@@ -1,0 +1,9 @@
+package blackjack.card;
+
+public enum Shape {
+
+    HEART,
+    SPADE,
+    CLOVER,
+    DIAMOND
+}
