--- src/main/java/card/Deck.java
@@ -1,36 +1,0 @@
-package card;
-
-import java.util.*;
-
-public class Deck {
-
-    private final Queue<Card> cards;
-
-    Deck(List<Card> cards) {
-        this.cards = new LinkedList<>(cards);
-    }
-
-    public static Deck createShuffledFullDeck() {
-        List<Card> cards = new LinkedList<>();
-        for (Shape shape : Shape.values()) {
-            cards.addAll(createNumberCardsOf(shape));
-        }
-        Collections.shuffle(cards);
-        return new Deck(cards);
-    }
-
-    private static List<Card> createNumberCardsOf(Shape shape) {
-        List<Card> cards = new ArrayList<>();
-        for (Number number : Number.values()) {
-            cards.add(new Card(shape, number));
-        }
-        return cards;
-    }
-
-    public Card draw() {
-        if (cards.isEmpty()) {
-            throw new IllegalStateException("[ERROR] 덱이 비어있습니다.");
-        }
-        return cards.poll();
-    }
-}
