--- src/main/java/card/Shape.java
@@ -1,9 +1,0 @@
-package card;
-
-public enum Shape {
-
-    HEART,
-    SPADE,
-    CLOVER,
-    DIAMOND
-}
